Add pitch flutter, completing KlattSynth
Three sine terms at fixed rates, each rebuilt from the fifty-entry quarter-wave
table across four quadrants, are summed and scaled first by pitch and then by
the flutter parameter, and the result is added to the pitch. That is what stops
a held vowel sounding mechanical.

fxmul_scaled now takes a 32-bit coefficient, because the second of those two
multiplies passes the flutter parameter straight out of the frame rather than
narrowing it first. Its body never depended on the narrower type.

All twenty-five functions in clsyn.obj are now transcribed, and every path
through KlattSynth that the tests can reach is bit-exact.

+++ src/klatt_fx.h
@@ -17,7 +17,7 @@ static inline int32_t mul32(int32_t a, int32_t b)
    so they pre-shift x by however much its magnitude demands and take the rest
    out of the final shift. The staged form drops low bits that a wider multiply
    would keep, so it cannot be folded back into a single expression. */
-static inline int32_t fxmul_scaled(int16_t coef, int32_t x)
+static inline int32_t fxmul_scaled(int32_t coef, int32_t x)
 {
     int32_t c = coef;
